refactor(kicad9): relocate overlap/power-bus/NC decisions to schematics/
Move find_overlapping_pins (was _find_wireable_nets), find_power_bus_runs
(decision half of _gen_power_bus_wires), and find_no_connect_pins (detection
half of _gen_no_connect_flags) into schematics/decisions.py. They take a
backend and return abstract render-mm data; sexp_schematic.py now calls them
through the Kicad9Backend adapter and emits via backend primitives. The KiCad
render geometry (_kicad_pin_pos, _render_xy) and Sexp construction stay in the
backend. esp32 parity unchanged (6 sheets/43 wires/37 labels/0 missing);
auto_stub suite 63 passed.

(cherry picked from commit 8466143bdfb1e12b8e5051c2bd8c25e10c71c856)
